module2-task1 - Добавляет сборку проекта и разбивку на модули
